Set TCP_USER_TIMEOUT in addition to keepalives
TCP keepalive probes only work when the connection is idle.  They're not
sent when there's unacked send data being retramnsmitted.  If the server
fails while we're retransmitting we don't break the connection and try
to elect and connect to a new server until the very long default
conneciton timeouts or the server comes back and the stale connection is
aborted.

We can set TCP_USER_TIMEOUT to break an unresponsive connection when
there's written data.  It changes the behavior of the keepalive probes
so we rework them a bit to clearly apply our timeout consistently
between the two mechanisms.

Signed-off-by: Zach Brown <zab@versity.com>